src/styleTransferService/requirements.txt
absl-py==0.9.0
astor==0.8.1
boto3==1.10.45
botocore==1.13.45
certifi==2019.11.28
docutils==0.15.2
gast==0.2.2
google-pasta==0.1.8
grpcio==1.26.0
h5py==2.10.0
jmespath==0.9.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
Markdown==3.1.1
numpy==1.18.0
opt-einsum==3.1.0
pika==1.1.0
Pillow==6.2.1
protobuf==3.11.2
python-dateutil==2.8.1
s3transfer==0.2.1
scipy==1.1.0
six==1.13.0
tensorboard==1.15.0
tensorflow==1.15.0
tensorflow-estimator==1.15.1
termcolor==1.1.0
urllib3==1.25.7
Werkzeug==0.16.0
wrapt==1.11.2
grpcio-tools==1.26.0
